Add backwards compatiblity verifying test for enum values. Move ResourceType and ResourceTypeVisitor to the db package.
