+++ src/test/java/bio/terra/janitor/db/ResourceTypeVisitorTest.java
@@ -1,13 +1,13 @@
-package bio.terra.janitor.app.common;
+package bio.terra.janitor.db;
 
-import static bio.terra.janitor.common.ResourceType.*;
+import static bio.terra.janitor.db.ResourceType.*;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import bio.terra.generated.model.*;
 import bio.terra.janitor.common.BaseUnitTest;
-import bio.terra.janitor.common.ResourceTypeVisitor;
 import bio.terra.janitor.common.exception.InvalidResourceUidException;
+import bio.terra.janitor.db.ResourceTypeVisitor;
 import org.junit.jupiter.api.Test;
 
 public class ResourceTypeVisitorTest extends BaseUnitTest {
